# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/overview:
    get:
      operationId: getAdminOverview
      tags: [admin]
      summary: Get admin dashboard overview stats
      responses:
        "200":
          description: Overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOverview"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all registered users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/documents:
    get:
      operationId: listAdminDocuments
      tags: [admin]
      summary: List knowledge base documents
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDocumentList"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/documents/{id}:
    post:
      operationId: updateAdminDocument
      tags: [admin]
      summary: Retry, activate, deactivate, or delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDocumentAction"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/documents/upload:
    post:
      operationId: uploadAdminDocument
      tags: [admin]
      summary: Upload a new knowledge base document
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DocumentUploadInput"
      responses:
        "200":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentUploadResult"
        "400":
          description: Invalid upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/training/upload:
    post:
      operationId: uploadAdminTrainingDataset
      tags: [admin]
      summary: Bulk upload a training dataset (Excel .xlsx or .csv)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/TrainingDatasetUploadInput"
      responses:
        "200":
          description: Dataset imported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingDatasetUploadResult"
        "400":
          description: Invalid upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/maintenance:
    get:
      operationId: getAdminMaintenance
      tags: [admin]
      summary: Get the current maintenance job status
      responses:
        "200":
          description: Maintenance status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMaintenanceStatus"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: startAdminMaintenance
      tags: [admin]
      summary: Start a maintenance job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminMaintenanceInput"
      responses:
        "200":
          description: Maintenance job start result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMaintenanceStartResult"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/training:
    get:
      operationId: listAdminTraining
      tags: [admin]
      summary: List recent training records
      responses:
        "200":
          description: Training records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTrainingList"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: addAdminTraining
      tags: [admin]
      summary: Add a training question/answer pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingRecordInput"
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    AdminOverview:
      type: object
      required:
        - documentsTotal
        - activeDocuments
        - usersTotal
        - adminsTotal
        - donationsTotal
        - donationsCount
        - trainingRowsTotal
        - recentUsers
        - recentDonations
      properties:
        documentsTotal:
          type: integer
        activeDocuments:
          type: integer
        usersTotal:
          type: integer
        adminsTotal:
          type: integer
        donationsTotal:
          type: integer
        donationsCount:
          type: integer
        trainingRowsTotal:
          type: integer
        recentUsers:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserSummary"
        recentDonations:
          type: array
          items:
            $ref: "#/components/schemas/AdminDonationSummary"
    AdminUserSummary:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        role:
          type: string
        createdAt:
          type: string
    AdminUserDetail:
      type: object
      required: [id, email, name, role, isActive, createdAt, chatCount, lastActiveAt]
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        chatCount:
          type: integer
        lastActiveAt:
          type: ["string", "null"]
    AdminUserList:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserDetail"
    AdminDonationSummary:
      type: object
      required: [eventId, amountCents, createdAt]
      properties:
        eventId:
          type: string
        amountCents:
          type: ["integer", "null"]
        createdAt:
          type: string
    AdminDocumentVersion:
      type: object
      required: [id, originalFilename, status, fileSizeBytes, uploadedAt]
      properties:
        id:
          type: string
        originalFilename:
          type: string
        status:
          type: string
        fileSizeBytes:
          type: integer
        uploadedAt:
          type: string
    AdminIngestionJob:
      type: object
      required: [id, status, progress, errorMessage]
      properties:
        id:
          type: string
        status:
          type: string
        progress:
          type: integer
        errorMessage:
          type: ["string", "null"]
    AdminDocument:
      type: object
      required: [id, title, status, createdAt, currentVersion, latestJob]
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
        createdAt:
          type: string
        currentVersion:
          oneOf:
            - $ref: "#/components/schemas/AdminDocumentVersion"
            - type: "null"
        latestJob:
          oneOf:
            - $ref: "#/components/schemas/AdminIngestionJob"
            - type: "null"
    AdminDocumentList:
      type: object
      required: [documents]
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/AdminDocument"
    AdminDocumentAction:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [retry, activate, deactivate, delete]
    DocumentUploadInput:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
    DocumentUploadResult:
      type: object
      required: [documentId, versionId, jobId, ingestionStatus]
      properties:
        documentId:
          type: string
        versionId:
          type: string
        jobId:
          type: string
        ingestionStatus:
          type: string
    AdminMaintenanceJob:
      type: object
      required: [id, action, status, requestedBy, startedAt, finishedAt, logs, error]
      properties:
        id:
          type: string
        action:
          type: string
        status:
          type: string
          enum: [running, succeeded, failed]
        requestedBy:
          type: string
        startedAt:
          type: string
        finishedAt:
          type: ["string", "null"]
        logs:
          type: array
          items:
            type: string
        error:
          type: ["string", "null"]
    AdminMaintenanceStatus:
      type: object
      required: [activeJob, history]
      properties:
        activeJob:
          oneOf:
            - $ref: "#/components/schemas/AdminMaintenanceJob"
            - type: "null"
        history:
          type: array
          items:
            $ref: "#/components/schemas/AdminMaintenanceJob"
    AdminMaintenanceInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [refresh-knowledge-index, refresh-training-dataset, refresh-all]
    AdminMaintenanceStartResult:
      type: object
      required: [started]
      properties:
        started:
          type: boolean
        activeJob:
          oneOf:
            - $ref: "#/components/schemas/AdminMaintenanceJob"
            - type: "null"
    TrainingRecord:
      type: object
      required: [id, sheet, row, question, answer]
      properties:
        id:
          type: integer
        sheet:
          type: string
        row:
          type: integer
        question:
          type: string
        answer:
          type: string
    AdminTrainingList:
      type: object
      required: [records, total]
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/TrainingRecord"
        total:
          type: integer
    TrainingRecordInput:
      type: object
      required: [question, answer]
      properties:
        question:
          type: string
        answer:
          type: string
    TrainingDatasetUploadInput:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
    TrainingDatasetUploadResult:
      type: object
      required: [added, skipped, total]
      properties:
        added:
          type: integer
        skipped:
          type: integer
        total:
          type: integer
...